+++ katarl2/envs/env_dmc.py
@@ -12,6 +12,7 @@ from katarl2.envs.common.env_cfg import EnvConfig
 
 @dataclass
 class DMCEnvConfig(EnvConfig):
+    max_episode_steps: int = 1000
     env_type: Literal['dmc'] = 'dmc'
     env_name: Literal[  # Only continuous control tasks
         # 20 tasks DMC_EASY_MEDIUM