+++ katarl2/envs/env_envpool.py
@@ -30,9 +30,10 @@ ENV_NAME = [
 ]
 
 # 并行版Wrapper, clearnl/ppo_atari_envpool.py (并修改为Gymnasium>=1.0版本)
-class RecordEpisodeStatistics:
-    def __init__(self, envs):
+class RecordEpisodeStatisticsAndTimeLimit:
+    def __init__(self, envs, time_limit: int):
         self.envs = envs
+        self.time_limit = time_limit
         self.num_envs = len(envs.all_env_ids)
         self.episode_returns = None
         self.episode_lengths = None
@@ -62,6 +63,10 @@ class RecordEpisodeStatistics:
         observations, rewards, terminations, truncations, infos = self.envs.step(action)
         self.episode_returns += infos["reward"]
         self.episode_lengths += 1
+        # 超时
+        infos["terminated"] |= self.episode_lengths > self.time_limit
+        truncations |= self.episode_lengths > self.time_limit
+
         self.returned_episode_returns[:] = self.episode_returns
         self.returned_episode_lengths[:] = self.episode_lengths
         # 当一局游戏结束是所有的生命值全部消耗完, lives=0或者terminated=True
@@ -75,7 +80,7 @@ class RecordEpisodeStatistics:
             'episode': episodic_info,
             '_episode': terminations | truncations
         }
-        infos['episodic_info'] = episodic_info  # 在强制中断游戏也可以获取当前episode的r,l
+        # infos['episodic_info'] = episodic_info  # 在强制中断游戏也可以获取当前episode的r,l
         if infos['terminated'].sum():
             infos['final_info'] = final_info
         return (
@@ -91,6 +96,7 @@ class RecordEpisodeStatistics:
 
 @dataclass
 class EnvpoolAtariConfig(EnvConfig):
+    max_episode_steps: int = 108000
     env_type: Literal['envpool'] = 'envpool'
     env_name: Literal[
         'Alien-v5', 'Amidar-v5', 'Assault-v5', 'Asterix-v5', 'Asteroids-v5',
@@ -114,15 +120,28 @@ def make_envpool_envs_from_cfg(cfg: EnvConfig, train: bool):
     if cfg.env_name in ATARI_NAME:
         # 默认支持 ClipReward, EpisodicLife, MaxAndSkip=4, FireReset
         num_envs = cfg.num_envs if train else cfg.num_eval_envs
+        # Reference: https://envpool.readthedocs.io/en/latest/env/atari.html
         envs = envpool.make(
             task_id=cfg.env_name,
-            env_type='gym',
+            env_type='gymnasium',
             num_envs=num_envs,
-            episodic_life=True,
-            reward_clip=True,
             seed=cfg.seed,
+            max_episode_steps=cfg.max_episode_steps,
+            repeat_action_probability=0.25,
+
+            # Default wrappers
+            noop_max=30,
+            stack_num=4,
+            episodic_life=True,
+            use_fire_reset=True,
+            reward_clip=True,
+            img_height=84,
+            img_width=84,
+            gray_scale=True,
+            frame_skip=4,
         )
     else:
         raise ValueError(f"Unsupported environment name: {cfg.env_name}. Supported names: {ATARI_NAME}")
-    envs = RecordEpisodeStatistics(envs)
+    # 这里的time_limit是经过frame_skip后的步数
+    envs = RecordEpisodeStatisticsAndTimeLimit(envs, time_limit=cfg.max_episode_steps // 4)
     return envs