+++ katarl2/envs/wrappers/time_limit.py
@@ -1,0 +1,23 @@
+import numpy as np
+import gymnasium as gym
+
+class TimeLimit(gym.Wrapper):
+    def __init__(self, env, duration):
+        super().__init__(env)
+        self._duration = duration
+        self._step = None
+
+    def step(self, action):
+        assert self._step is not None, "Must reset environment."
+        obs, reward, done, info = self.env.step(action)
+        self._step += 1
+        if self._step >= self._duration:
+            done = True
+            if "discount" not in info:
+                info["discount"] = np.array(1.0).astype(np.float32)
+            self._step = None
+        return obs, reward, done, info
+
+    def reset(self):
+        self._step = 0
+        return self.env.reset()